Remove inits and pthread from gf
The global arrays can be initialized at compile time.  Therefore, the init
functions are obsolete and can be replaced by statically initialized constant
arrays (which depend only on the prim polynomial) which is fixed at compile time
as well. Since there is no reason to dynamically update the polynomial at
runtime this is perfectly ok.

The gfgenerate function generates the header files consisting of the necessary
tables based on the chosen polynomial.